# doc/api.yaml
openapi: 3.0.3
servers:
  - url: http://127.0.0.1:8080/
info:
  title: WasaPhoto
  description: |-
    This is the project for Wasa (web and software
    architecture)
  version: "b-1.1.0"
paths:
...
  /addphoto:
    post:
      tags: ["addPhoto"]
      operationId: addPhoto
      summary: Post a new photo
      description: |
        Post a new photo
      security:
        - LegacyAuth: []
      requestBody:
        description: |
          The user must provides the username, id and
          the photo data
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                Username:
                  type: string
                  example: Maria
                  pattern: 'ˆ.*?$'
                  minLength: 3
                  maxLength: 16
                Id:
                  type: string
                  example: a35f6dfa-69dc-11ed-a1eb-0242ac120002
                  pattern: '^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$'
                Text:
                  type: string
                  example: Test
                  pattern: 'ˆ.*?$'
                  minLength: 1
                  maxLength: 256
                myFile:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: Photo uploaded successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/basicProfile"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /search:
    parameters:
      - $ref: "#components/parameters/Search"
    get:
      tags: ["searchProfile"]
      operationId: searchProfile
      summary: Search fo User
      description: |-
        Search by query a user.
        It'll always return a list of ultra-basic
        information of the user.
      responses:
        "200":
          description: Status Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ultraBasicProfile"
components:
  schemas:
    basicProfile:
      title: Get Basic Profile Info
      type: object
      properties:
        Username:
          type: string
          example: Maria
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
        Id:
          type: string
          example: a35f6dfa-69dc-11ed-a1eb-0242ac120002
          pattern: '^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$'
        ProfilePictureLocation:
          type: string
          example: /images/a35f6dfa-69dc-11ed-a1eb-0242ac120002.png
          pattern: '^/images/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}.png$'
        Name:
          type: string
          example: Mario
          maxLength: 20
          minLength: 0
        Surname:
          type: string
          example: Rossi
          maxLength: 20
          minLength: 0
        Followers:
          type: integer
          minLength: 0
        Followings:
          type: integer
          minimum: 0
        Images:
          type: array
          items:
            $ref: "#/components/schemas/basicImage"
    basicImage:
      title: Get Basic Image Info
      type: object
      properties:
        idImage:
          type: string
          example: a35f6dfa-69dc-11ed-a1eb-0242ac120002
          pattern: '^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$'
        Location:
          type: string
          example: /images/a35f6dfa-69dc-11ed-a1eb-0242ac120002.png
          pattern: '^/images/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}.png$'
        Time:
          type: string
        Likes:
          type: integer
          minimum: 0
        Comments:
          type: integer
          minimum: 0
        Text:
          type: string
          example: Text Image
          minLength: 1
          maxLength: 256
...
    ultraBasicProfile:
      title: Get Ultra Basic Profile Info
      type: object
      properties:
        Username:
          type: string
          pattern: 'ˆ.*?$'
          minLength: 3
          maxLength: 16
        Id:
          type: string
          example: a35f6dfa-69dc-11ed-a1eb-0242ac120002
          pattern: '^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$'
        ProfilePictureLocation:
          type: string
          example: /images/a35f6dfa-69dc-11ed-a1eb-0242ac120002.png
          pattern: '^/images/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}.png$'
...
  securitySchemes:
    LegacyAuth:
      type: apiKey
      in: header
      name: X-App-Token
      description: |-
        App session token value. This value is sent by the server in the login
        and session refresh APIs.
  responses:
    Unauthorized:
      description: The access token is missing or it isn't matchin with username
      headers:
        WWW_Authenticate:
          schema:
            type: string
    BadRequest:
      description: Missing fields
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
...